+++ CHANGELOG.md
@@ -15,6 +15,8 @@ These changes are available on the `master` branch, but have not yet been releas
 - Add `RoleColours.HOLOGRAPHIC_PRIMARY`, `RoleColours.HOLOGRAPHIC_SECONDARY`, and
   `RoleColours.HOLOGRAPHIC_TERTIARY` class constants.
   ([#3268](https://github.com/Pycord-Development/pycord/pull/3268))
+- Add `Asset.size` and `Asset.extension` properties.
+  ([#3294](https://github.com/Pycord-Development/pycord/pull/3294))
 
 ### Changed
 
@@ -54,6 +56,8 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#3278](https://github.com/Pycord-Development/pycord/pull/3278))
 - Fix `PartialMessage.edit()` to work with `DesignerView`.
   ([#3237](https://github.com/Pycord-Development/pycord/pull/3237))
+- Fix animated asset url extension from `.gif` to `.webp`
+  ([#3294](https://github.com/Pycord-Development/pycord/pull/3294))
 
 ### Deprecated